python/requirements.txt

cffi>=1.9.1
pyzmq>=15.3.0
six>=1.10.0
